Added `ha-form-mgc-list` for use in `ha-form` schema Color thresholds and state map now use the new `ha-form-mgc-list` Removed old unused components Moved schemas to `editorConst.js` Translations are now handled through `hass.localize` Color selector should now display proper color when supplied with color name Fixed inability to clear few options Removed leftover comments
